Insights: Refactor backend code for adding more code
