+++ server/lib/insights/detector.py
